pyproject.toml
[tool.poetry]
name = "rivian_exporter"
version = "0.1.0"
description = "Prometheus exporter for Rivian"
authors = ["John Oxley <john.oxley@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "rivian_exporter", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
glog = "^0.3.1"
prometheus-client = "^0.17.1"
rivian-python-client = "^1.0.4"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
pytest = '^7.2.2'
pytest-asyncio = "^0.18.3"
testslide = "^2.7.1"
black = "^23.9.1"
isort = "^5.12.0"
coverage = { extras = ["toml"], version = "^7.3.2" }
flake8 = "^6.1.0"
flake8-bugbear = "^23.9.16"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [".*", "*/site-packages/*"]

[tool.coverage.report]
fail_under = 80

[tool.mypy]
warn_return_any = true

[[tool.mypy.overrides]]
module = "glog"
ignore_missing_imports = true
